SpreadsheetSelectionHistoryToken.parse0 pulled by subclasses
- parts of the parse0 method were extracted by Cell/Column/Row/LabelMapping
